requirements.txt
cryptography
pysocks
requests
stem